work around ojdbc quirks
Oracle sometimes closes a cursor and its result set on the server without notification to the OJDBC driver when all rows have been processed. This results in an unexpected ORA-17010 exception thrown when calling next(). This only happens when multiple cursors/result sets are used simultaneously, especially in multi-threaded applications. The fix swallows the ORA-17ß1ß exception if it happens in ResultSet.next(). Tested with ojdbc17.